Revert incomplete changes to serving rce-api via 443
And serve via HTTPS on port 3000.

Serving the RCE-API via port 443 (rather than an alternate port;
3000 in out case) is appealing as it reduces issues with CORS and means
one less port open. However, the fact that it requires changes to
apps.js & app/application.js means it's a litte hacky IMO.

We can discuss further and for now I have pushed the required changes
to a separate branch in my personal repo in case we decide to run with
that as some future point. It can be found here:

https://github.com/JedMeister/canvas/tree/serve-rce-api-on-443